Move cart list page component (#178)
